refactor: collapse repository + client into single KnowledgeGraphClient
Removes the KnowledgeGraphApiRepository/StandardsRepository abstraction layer.
KnowledgeGraphClient now handles HTTP calls directly alongside its caching
and concurrency logic — one class, one responsibility.

- repository.ts removed; HTTP logic moved into client.ts private methods
- StandardsRepository interface removed — no longer needed without the
  JSON repository variant
- Tests consolidated into client.test.ts covering all HTTP paths,
  caching eviction, pagination, and error mapping
