+++ sdks/typescript/src/knowledge-graph/index.ts
@@ -1,4 +1,3 @@
-export type { AcademicStandard, LearningComponent, StandardInfo, StandardsRepository } from './types.js';
+export type { AcademicStandard, LearningComponent, StandardInfo } from './types.js';
 export { parseGradeFromStandard } from './types.js';
-export { KnowledgeGraphApiRepository } from './repository.js';
 export { KnowledgeGraphClient } from './client.js';
